# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/query": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Run Query",
                "parameters": [
                    {
                        "description": "Query Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.runQuery.queryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        },
        "/api/user/sync": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sync Strava data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Strava Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        },
        "/api/user/task/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get Task Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.runQuery.queryInput": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
...
        "queue.TaskResult": {
            "type": "object",
            "properties": {
                "completionTime": {
                    "type": "string"
                },
                "error": {},
                "id": {
                    "type": "string"
                },
                "result": {},
                "status": {
                    "$ref": "#/definitions/queue.TaskStatus"
                }
            }
        },
        "queue.TaskStatus": {
            "type": "string",
            "enum": [
                "pending",
                "success",
                "error",
                "archived"
            ],
            "x-enum-varnames": [
                "Pending",
                "Success",
                "Error",
                "Archived"
            ]
        },
...
    }
}